cleaned up files for a smoother review 🍻
