+++ lib/main.dart
@@ -3,7 +3,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:niira/screens/lobby.dart';
-import 'package:niira/screens/waiting_for_game_to_start.dart';
+import 'package:niira/screens/waiting_screen/waiting_for_game_to_start.dart';
 import 'package:niira/screens/welcome.dart';
 import 'package:niira/services/auth/auth_service.dart';
 import 'package:niira/services/auth/firebase_auth_service.dart';
@@ -61,9 +61,7 @@ class MyApp extends StatelessWidget {
             stream: _authService.streamOfAuthState,
             builder: (context, snapshot) {
               // TODO: check for snapshot error and send to navigation manager for display
-              return (snapshot.data == null)
-                  ? WelcomeScreen()
-                  : WaitingForGameToStartScreen();
+              return (snapshot.data == null) ? WelcomeScreen() : LobbyScreen();
             },
           )),
     );
