+++ lib/screens/waiting_screen/waiting_for_game_to_start.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:niira/models/game.dart';
 import 'package:niira/models/player.dart';
 import 'package:niira/screens/lobby.dart';
+import 'package:niira/screens/waiting_screen/joined_players_list.dart';
 import 'package:niira/services/database/database_service.dart';
 import 'package:niira/services/navigation_service.dart';
 import 'package:provider/provider.dart';
@@ -19,7 +20,7 @@ class _WaitingForGameToStartScreenState
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      key: Key('c'),
+      key: Key('waiting_for_game_to_start_screen'),
       appBar: AppBar(
         automaticallyImplyLeading: false,
         title: Text('Waiting for game to start'),
@@ -64,75 +65,6 @@ class _WaitingForGameToStartScreenState
               }
             }),
       ),
-    );
-  }
-}
-
-// show list of all joined players to all joined players
-class JoinedPlayersList extends StatelessWidget {
-  final List<Player> joinedPlayers;
-
-  JoinedPlayersList({
-    @required this.joinedPlayers,
-    Key key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: EdgeInsets.fromLTRB(0, 10, 0, 0),
-      child: ListView.builder(
-        itemCount: joinedPlayers.length,
-        itemBuilder: (context, index) => JoinedPlayerTile(
-          player: joinedPlayers[index],
-        ),
-      ),
-    );
-  }
-}
-
-// show joined player including their name and indicating if they have been chosen as tagger
-class JoinedPlayerTile extends StatelessWidget {
-  JoinedPlayerTile({
-    @required this.player,
-  });
-
-  final Player player;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      key: Key('created_game_tile_${player.id}'),
-      padding: EdgeInsets.fromLTRB(20, 5, 20, 5),
-      child: player.isTagger
-          ? Card(
-              key: Key('tagger_tile_${player.id}'),
-              color: Theme.of(context).accentColor,
-              child: ListTile(
-                title: Text(
-                  player.username,
-                  style: TextStyle(
-                    fontSize: 22,
-                    color: Colors.white,
-                  ),
-                ),
-                subtitle: Text(
-                  'is the tagger',
-                  style: TextStyle(color: Colors.white),
-                ),
-              ),
-            )
-          : Card(
-              child: ListTile(
-                title: Text(
-                  player.username,
-                  style: TextStyle(
-                    fontSize: 22,
-                  ),
-                ),
-                subtitle: Text('has joined'),
-              ),
-            ),
     );
   }
 }
